Rename `_pmi.py` -> `pmi.py` (#115)
