+++ soynlp/word/pmi.py
